Specify an extension for tunnelling external protocols
There are various extensions for extending the RFB protocol beyond
simply handling its original framebuffer and input device features.
For example, to extend it to handle audio, or file transfer.

It is possible to use standalone protocols and deploy their servers
alongside the VNC server, however, that having them all integrated
with RFB is very desirable from a managability and usability POV
because you get consistent authentication and encryption across
all features, and concepts such as tunnelling RFB over HTTP
WebSockets just work. Integrating the features in RFB has downsides
in terms of re-inventing the wheel vs extisting use case specific
protocols, and has challenges in terms of latency. eg you don't
want a large framebuffer update to glitch your audio or USB
traffic.

The SPICE project/protocol took the approach of using a multi
connection model to address this. The initial connection to the
SPICE server servers has the coordinator and allows negotiation
of the use of further connections to handle each different type
of data.

This spec addition introduces a similar concept to RFB by
defining a protocol extension for using RFB to "tunnel" access
to external protocols.  In this mode the initial RFB security
type negotiation happens as normal, but after SetEncodings is
run, the client can opt to switch a connection over to run an
external protocol. This protocol is thus automatically protected
by the RFB authentication that has just taken place, including
any session encryption layer (ie TLS) that it negotiated the
use of.

This extension is desired for QEMU to enable its VNC server
to achieve closer feature parity with SPICE:

  - USB redirection - QEMU has a USB redirector that allows
    a SPICE client to expose a local USB device to the remote
    server. This USB redir proocol has no built-in strong
    authentication or encryption facilities of its own, so
    relies on those provided by an encapsulating transport
  - Character device - the (usually human targetted) byte
    stream associated with a serial port or a virutal console
    device
  - NBD - a way to expose a locally stored ISO image to a
    remote server for use as media in an emulated CDROM

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>